pe-coupling-analyzer/requirements.txt
numpy>=1.26.0,<2.0
pandas>=2.2.0
scipy>=1.12.0
onnxruntime>=1.17.0
opencv-python>=4.9.0
PySide6>=6.6.0
python-docx>=1.1.0
pywt>=1.5.0
matplotlib>=3.8.0
h5py>=3.10.0
pyyaml>=6.0.1
pyinstaller>=6.4.0
pytest>=8.0.0
